Renombramos la clase Map a GameMap para evitar conflictos con el Map de Java. Renombrabamos el Package mapa a gamemap
