+++ src/main/java/com/chucknorris/gamemap/YellowNode.java
@@ -1,24 +1,24 @@
-package com.chucknorris.mapa;
+package com.chucknorris.gamemap;
 
 import com.chucknorris.commons.Position;
 import com.chucknorris.player.Player;
 
 import java.util.ArrayList;
 
-public class RedNode extends Node {
+public class YellowNode extends Node {
 
-    public RedNode(ArrayList<Node> next, Position pos) {
+    public YellowNode(ArrayList<Node> next, Position pos) {
         super(next, pos);
     }
 
     @Override
     public void applyRewards(Player p) {
-        p.addCoins(3);
+        p.addCoins(10);
     }
 
     @Override
     public String getType() {
-        return "RED";
+        return "YELLOW";
     }
 
 }
